# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LaraSol API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/wards": {
            "get": {
                "tags": [
                    "Wards"
                ],
                "description": "Get all Wards",
                "operationId": "getWardsList",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Wards"
                ],
                "summary": "Store new Ward",
                "description": "Returns ward data",
                "operationId": "storeWards",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreWardRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Wards"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/wards/{id}": {
            "get": {
                "tags": [
                    "Wards"
                ],
                "summary": "Get ward information",
                "description": "Returns ward data",
                "operationId": "getWardById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Ward id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Wards"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "Wards"
                ],
                "summary": "Update existing ward",
                "description": "Returns updated ward data",
                "operationId": "updateWards",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Ward id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateWardRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Wards"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Wards"
                ],
                "summary": "Delete existing ward",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteWard",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Ward id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "StoreWardRequest": {
                "title": "Ward Form store request",
                "required": [
                    "ward_name"
                ],
                "properties": {
                    "ward_id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "ward_name": {
                        "description": "Ward Name",
                        "type": "string",
                        "example": "Starehe"
                    },
                    "county_id": {
                        "type": "integer",
                        "example": "1"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "StoreWardRequest"
                }
            },
...
            "UpdateWardRequest": {
                "title": "Ward Form request",
                "required": [
                    "ward_name"
                ],
                "properties": {
                    "ward_id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "ward_name": {
                        "description": "Ward Name",
                        "type": "string",
                        "example": "Starehe"
                    },
                    "county_id": {
                        "type": "integer",
                        "example": "1"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "UpdateWardRequest"
                }
            },
...
            "Wards": {
                "required": [
                    "ward_name",
                    "county_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "ward_name": {
                        "description": "Ward name",
                        "type": "string",
                        "readOnly": "true"
                    },
                    "county_id": {
                        "description": "County id",
                        "type": "integer",
                        "readOnly": "true"
                    },
                    "created_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/created_at"
                    },
                    "updated_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/updated_at"
                    },
                    "deleted_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/deleted_at"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Wards"
                }
            }
        }
    }
}